Return the average count

+++ app/graphql/types/events/history_stat.rb
@@ -9,7 +9,7 @@ module Types
       field :name, String, null: false
       field :type, Types::Events::HistoryType, null: false
       field :count, Integer, null: false
-      field :average_events_per_visitor, Integer, null: false
+      field :average_events_per_visitor, Float, null: false
     end
   end
 end
